+++ app/templates/base.html
@@ -33,7 +33,7 @@
                             <a class="nav-link" href="{{ url_for('login.login') }}">Login</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('register') }}">Register</a>
+                            <a class="nav-link" href="{{ url_for('register.register') }}">Register</a>
                         </li>
                     {% endif %}
                 </ul>
