+++ app/templates/login.html
@@ -13,5 +13,5 @@
     </div>
     <button type="submit" class="btn btn-primary">Login</button>
 </form>
-<a href="{{ url_for('register') }}">Don't have an account? Register here.</a>
+<a href="{{ url_for('register.register') }}">Don't have an account? Register here.</a>
 {% endblock %}
