+++ app/templates/register.html
@@ -13,5 +13,5 @@
     </div>
     <button type="submit" class="btn btn-primary">Register</button>
 </form>
-<a href="{{ url_for('login') }}">Already have an account? Login here.</a>
+<a href="{{ url_for('login.login') }}">Already have an account? Login here.</a>
 {% endblock %}
